requirements.txt
# Core
Django==5.2.4
sc4py==0.1.4

# Models
django-better-choices==1.17
django-polymorphic==4.1.0

# Rule engine
jsonschema==4.24.0
rule-engine==4.5.3

# REST
# django-ninja==1.4.1
requests==2.32.4

# Session
django-redis-sessions==0.6.2
redis==6.2.0
#valkey[libvalkey]==6.1.0

# # Security
# django-oauth-toolkit==3.0.1
# cryptography==45.0.5
# pyjwt==2.10.1
# django-cors-headers==4.7.0


# Audit
django-simple-history==3.10.1
# django-safedelete==1.4.1

# UI
# django-admin-kubi
# django-daisy==1.0.23
# django-unfold==0.62.0
# django-theme-adminlte3==3.2.0.12
django-json-widget==2.0.1
django-import-export==4.3.8
# django-compressor==4.5.1
# django-sass-processor==1.4.1
django-richtextfield==1.6.2
# django-daterange-filter==1.3.0
# django-admin-autocomplete-filter==0.7.1
whitenoise==6.9.0


# Database
psycopg[binary,pool]==3.2.9

# # Image
# pillow==11.3.0

# Production
sentry-sdk[django,rq]==2.32.0
gunicorn==23.0.0
